Implement Authentication and Authorization for DELETE endpoint

+++ app/api.py
@@ -1,6 +1,6 @@
-from fastapi import APIRouter, Query, Path, HTTPException, status
-from .schemas import SearchField, SearchResponse, StudentCreate, StudentCreatePatch
-from . import services
+from fastapi import APIRouter, Query, Path, HTTPException, status, Depends
+from .schemas import SearchField, SearchResponse, StudentCreate, StudentCreatePatch, UserLogin
+from . import services, auth
 
 router = APIRouter(prefix="/student_marks", tags=["Students"])
 
@@ -14,7 +14,7 @@ def search_students(
     student_marks = services.load_students_marks_data()
     all_students = list(student_marks)
     
-    try:
+    try:    
         filtered_students = services.filter_students_marks_data_by_field(all_students, search_field, search_value)
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -63,11 +63,40 @@ def update_student_patch(student_id: int = Path(..., ge=1), student_data: Studen
         "student": updated_student_data
     }
 
-@router.delete("/delete/{student_id}")
-def delete_student(student_id: int = Path(..., ge=1)):
-    deleted_student_data = services.delete_student_data(student_id)
+
+
+@router.post("/login", tags=["Security"])
+def login_for_access_token(user_data: UserLogin):
+    users = auth.load_users()
+
+    
+    if user_data.username not in users or users[user_data.username]["password"] != user_data.password:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password"
+        )
+
+    # Generate a fake token for demonstration
+    fake_token = f"{user_data.username}-secret-token"
+    
     return {
-        "message": "Student deleted successfully",
-        "student": deleted_student_data
+        "access_token": fake_token, 
+        "token_type": "bearer"
     }
 
+
+#--------------------Secure DELETE Logic---------------------
+@router.delete("/{student_id}")
+def remove_student(student_id: int, current_user: dict = Depends(auth.get_current_user)):
+
+    if current_user["role"] != "admin":
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only admins can delete students"
+        )
+    
+    try:
+        result = services.delete_student_data(student_id)
+        return result
+    except ValueError as e:
+        raise HTTPException(status_code=404, detail=str(e))
